+++ src/modules/products/products.service.ts
@@ -1,4 +1,4 @@
-import { Injectable } from '@nestjs/common';
+import { Injectable, NotFoundException } from '@nestjs/common';
 import { PrismaService } from 'src/prisma/prisma.service';
 import { MovementsService } from '../movements/movements.service';
 import { CreateProductDto } from './dto/create-product.dto';
@@ -11,9 +11,12 @@ export class ProductsService {
     private movementsService: MovementsService,
   ) {}
 
-  async create(data: CreateProductDto) {
+  async create(data: CreateProductDto, userId: number) {
     const product = await this.prisma.product.create({
       data: {
+        user: {
+          connect: { id: userId },
+        },
         name: data.name,
         description: data.description,
         quantity: data.quantity,
@@ -23,6 +26,7 @@ export class ProductsService {
 
     if (data.quantity > 0) {
       await this.movementsService.create({
+        userId,
         productId: product.id,
         quantity: data.quantity,
         type: 'entrada',
@@ -32,27 +36,48 @@ export class ProductsService {
     return product;
   }
 
-  findAll() {
-    return this.prisma.product.findMany();
+  findAll(userId: number) {
+    return this.prisma.product.findMany({
+      where: { userId },
+      include: { movements: true },
+    });
   }
 
-  findOne(id: number) {
-    return this.prisma.product.findUnique({ where: { id } });
+  findOne(id: number, userId: number) {
+    return this.prisma.product.findFirst({
+      where: { id, userId },
+    });
   }
 
-  update(id: number, updateProductDto: UpdateProductDto) {
-    return this.prisma.product.update({
-      where: { id },
+  update(id: number, updateProductDto: UpdateProductDto, userId: number) {
+    return this.prisma.product.updateMany({
+      where: { id, userId },
       data: updateProductDto,
     });
   }
 
-  remove(id: number) {
-    return this.prisma.product.delete({ where: { id } });
+  async remove(id: number, userId: number) {
+    const product = await this.prisma.product.findFirst({
+      where: { id, userId },
+    });
+
+    if (!product) {
+      throw new NotFoundException(
+        `Produto ${id} não encontrado para o usuário.`,
+      );
+    }
+
+    await this.prisma.movement.deleteMany({
+      where: { productId: id, userId },
+    });
+
+    return this.prisma.product.delete({
+      where: { id, userId },
+    });
   }
 
-  async removeAll() {
-    await this.prisma.movement.deleteMany();
-    return this.prisma.product.deleteMany();
+  async removeAll(userId: number) {
+    await this.prisma.movement.deleteMany({ where: { userId } });
+    return this.prisma.product.deleteMany({ where: { userId } });
   }
 }
